improve regulation structure
