+++ app/regulations/v1/undergraduate_regulation.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module V1
-  class UndergraduateRegulation < Regulation::Base
-    identifier :undergraduate
-    name       'Lisans Yönetmeliği'
-    version    31_103
+  class UndergraduateRegulation < Extensions::Regulation::Base
+    identifier      :undergraduate
+    number          31_103
+    effective_date '23/05/2020'
   end
 end
